First piece of package building.
I'm attempting to build a way for gulp to crush a "package" of JS and
CSS files together into an HTML template. This is to give more control
to the "app" part of the build process so it can selectively include
specific JS or CSS files.

I'm also hoping to raise the code quality of the gulp file because the
one I had before has logic that's pretty hard to follow.

Note that the `build/` directory is temporarily being included in the
git repo. If this stops being a play thing, the source files this gulp
file is using will be generated by the "app" part of the build.

Test Plan:
Run `gulp crush` from the `sandbox/` directory, and observe that two
files are created in `build/crush-output/`.