Added serve target in make.
It's a fairly underwhelming way to serve files, but will keep me happy
for a little while longer. I'm definitely going to want an auto-rebuild
setup at some point. And bringing down the SimpleHTTPServer that Python
provides is always a pain.

Test Plan:
Run `make build serve` and then go to the URL it prints out. See that
the pages are being served correctly.