Made nav bar more consistent across pages.
My thinking previously was that the top bar of the post pages, because
it contains breadcrumbs rather than a flat list of links, should
receive a different visual treatment.

I'm not sure this isn't true, but the way things were before made it
hard to see that the top bit was different from the content. There
wasn't really anything that made it seem like a top bar...

Test Plan:
`make build serve` and go to one of the posts and see that it has the
same blue borders as the other pages.