Separate post metadata from page metadata.
I didn't realize NextJS was using the post metadata to fill out the
meta tags. This was doing a couple things: setting the title to post
title (yay!) and setting the meta description to the description (not
yay).

The description is "not yay" because the description is often a React
object, so the meta description would be literally "[object Object]".

I've fixed this with a layer of indirection: `createMetadata`. It will
create the NextJS metadata for the page and tuck the actual post
metadata off to the side of the object. This lets me access it easily
and do things like modify the page meta tags based on the post
metadata.